refactor(protocol): abstract gossipsub patches into own module
